deployments/fs-mcp/pyproject.toml
[project]
name = "bash-mcp"
version = "0.1.0"
description = "Bash MCP Server with data analysis libraries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.3.0",
    "pandas>=2.2.0",
    "scipy>=1.15.0",
    "numpy>=2.2.0",
    "mplfinance>=0.12.10",
    "lightweight-charts>=2.1.0",
    "psutil>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
